Moved docker-compose.yml for Sail to root directory
The docker-compose.yml file needs to be in the root directory in order for the .env file variables to be read correctly. The "sail" command itself will read the .env variables correctly, however when it comes to Docker, the MariaDB service in particular doesn't pickup the environment variables. Setting the "env_file" setting in YAML explicitly doesn't work because it only takes the .env variables and doesn't merge them with the "environment" variables. The only alternative is passing the "--env-file" option to the sail command every time.