Created seperate presenter layer
The presenter layer allows for functionality related to a model and the presentation to be re-used. This is for generating data from an existing model, without modifying the model. An example is generating a URL for a model, there's instances where a URL needs to be generated outside of a view and it shouldn't just be in the model.